--- a/aggregationpb/aggregation_vtproto.pb.go
+++ b/aggregationpb/aggregation_vtproto.pb.go
@@ -1352,13 +1352,16 @@ var vtprotoPool_CombinedMetrics = sync.Pool{
 }
 
 func (m *CombinedMetrics) ResetVT() {
-	for _, mm := range m.ServiceMetrics {
-		mm.ResetVT()
+	for k, mm := range m.ServiceMetrics {
+		mm.ReturnToVTPool()
+		m.ServiceMetrics[k] = nil
 	}
+	f0 := m.ServiceMetrics[:0]
 	m.OverflowServices.ReturnToVTPool()
-	f0 := m.OverflowServiceInstancesEstimator[:0]
+	f1 := m.OverflowServiceInstancesEstimator[:0]
 	m.Reset()
-	m.OverflowServiceInstancesEstimator = f0
+	m.ServiceMetrics = f0
+	m.OverflowServiceInstancesEstimator = f1
 }
 func (m *CombinedMetrics) ReturnToVTPool() {
 	if m != nil {
@@ -1419,11 +1422,14 @@ var vtprotoPool_ServiceMetrics = sync.Pool{
 }
 
 func (m *ServiceMetrics) ResetVT() {
-	for _, mm := range m.ServiceInstanceMetrics {
-		mm.ResetVT()
+	for k, mm := range m.ServiceInstanceMetrics {
+		mm.ReturnToVTPool()
+		m.ServiceInstanceMetrics[k] = nil
 	}
+	f0 := m.ServiceInstanceMetrics[:0]
 	m.OverflowGroups.ReturnToVTPool()
 	m.Reset()
+	m.ServiceInstanceMetrics = f0
 }
 func (m *ServiceMetrics) ReturnToVTPool() {
 	if m != nil {
@@ -1463,16 +1469,25 @@ var vtprotoPool_ServiceInstanceMetrics = sync.Pool{
 }
 
 func (m *ServiceInstanceMetrics) ResetVT() {
-	for _, mm := range m.TransactionMetrics {
-		mm.ResetVT()
+	for k, mm := range m.TransactionMetrics {
+		mm.ReturnToVTPool()
+		m.TransactionMetrics[k] = nil
 	}
-	for _, mm := range m.ServiceTransactionMetrics {
-		mm.ResetVT()
+	f0 := m.TransactionMetrics[:0]
+	for k, mm := range m.ServiceTransactionMetrics {
+		mm.ReturnToVTPool()
+		m.ServiceTransactionMetrics[k] = nil
 	}
-	for _, mm := range m.SpanMetrics {
-		mm.ResetVT()
+	f1 := m.ServiceTransactionMetrics[:0]
+	for k, mm := range m.SpanMetrics {
+		mm.ReturnToVTPool()
+		m.SpanMetrics[k] = nil
 	}
+	f2 := m.SpanMetrics[:0]
 	m.Reset()
+	m.TransactionMetrics = f0
+	m.ServiceTransactionMetrics = f1
+	m.SpanMetrics = f2
 }
 func (m *ServiceInstanceMetrics) ReturnToVTPool() {
 	if m != nil {