See #16. Use vtproto fork to generate code that does not have the pooling bug.

The changes to vtproto: https://github.com/planetscale/vtprotobuf/compare/main...carsonip:vtprotobuf:fix-pooling
It fixes our case and generates the right code, but not sure if it breaks any code outside our use case.

benchstat:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-aggregation/aggregators
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                            │ bench-out/vtproto-fix-codegen-before │ bench-out/vtproto-fix-codegen-after │
                            │                sec/op                │    sec/op     vs base               │
AggregateCombinedMetrics-16                           3.537µ ±  3%    3.416µ ± 5%        ~ (p=0.240 n=6)
AggregateBatchSerial-16                               18.02µ ±  4%    17.55µ ± 3%   -2.60% (p=0.041 n=6)
AggregateBatchParallel-16                             19.69µ ±  6%    18.90µ ± 6%   -4.01% (p=0.015 n=6)
CombinedMetricsEncoding-16                            8.241µ ±  2%    5.974µ ± 1%  -27.51% (p=0.002 n=6)
CombinedMetricsDecoding-16                            4.268µ ±  1%    4.295µ ± 2%        ~ (p=0.485 n=6)
CombinedMetricsToBatch-16                             426.0µ ±  3%    414.6µ ± 2%   -2.66% (p=0.009 n=6)
EventToCombinedMetrics-16                             1.022µ ± 10%    1.034µ ± 4%        ~ (p=0.513 n=6)
geomean                                               10.98µ          10.32µ        -6.00%

                            │ bench-out/vtproto-fix-codegen-before │ bench-out/vtproto-fix-codegen-after │
                            │                 B/op                 │     B/op      vs base               │
AggregateCombinedMetrics-16                          3.606Ki ± 12%   3.477Ki ± 1%   -3.57% (p=0.002 n=6)
AggregateBatchSerial-16                              22.72Ki ±  1%   22.19Ki ± 1%   -2.34% (p=0.002 n=6)
AggregateBatchParallel-16                            22.64Ki ±  1%   22.29Ki ± 1%   -1.54% (p=0.002 n=6)
CombinedMetricsEncoding-16                            2624.0 ±  0%     576.0 ± 0%  -78.05% (p=0.002 n=6)
CombinedMetricsDecoding-16                           10.01Ki ±  0%   10.01Ki ± 0%        ~ (p=0.182 n=6)
geomean                                              8.620Ki         6.269Ki       -27.27%

                            │ bench-out/vtproto-fix-codegen-before │ bench-out/vtproto-fix-codegen-after │
                            │              allocs/op               │ allocs/op   vs base                 │
AggregateCombinedMetrics-16                             41.00 ± 7%   34.00 ± 3%  -17.07% (p=0.002 n=6)
AggregateBatchSerial-16                                 182.5 ± 1%   159.0 ± 1%  -12.88% (p=0.002 n=6)
AggregateBatchParallel-16                               183.0 ± 0%   159.0 ± 0%  -13.11% (p=0.002 n=6)
CombinedMetricsEncoding-16                              77.00 ± 0%   45.00 ± 0%  -41.56% (p=0.002 n=6)
CombinedMetricsDecoding-16                              40.00 ± 0%   40.00 ± 0%        ~ (p=1.000 n=6) ¹
geomean                                                 84.14        68.85       -18.17%
¹ all samples are equal
```